% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_na_level}
\alias{add_na_level}
\title{For factor columns with NAs, convert NA to its own level}
\usage{
add_na_level(X)
}
\description{
For factor columns with NAs, convert NA to its own level
}

