% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{caret_predict}
\alias{caret_predict}
\title{Make predictions using a trained caret model}
\usage{
caret_predict(trained_model, newdata)
}
\arguments{
\item{trained_model}{The output of caret_train.}

\item{newdata}{A data_list of the same form as that input to caret_train.}
}
\value{
y_hat for the new data.
}
\description{
Make predictions using a trained caret model
}

