% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{caret_train}
\alias{caret_train}
\title{Train a generic caret model on the available data sets}
\usage{
caret_train(data_list, opts = list())
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
  train: A data.frame giving the artist, track, user, rating, and time info. \cr
  words: A matrix giving word indicators for artist-user pairs. \cr
  users: A matrix givin gsurvey results for each user. \cr}

\item{opts}{A list of options specifyign training. See merge_caret_opts() for
options and their defaults.}
}
\value{
A list with the following elements,
  $opts A list of options passed into training.
  $model The full trained caret model output.
}
\description{
Train a generic caret model on the available data sets
}
\examples{
data(train)
data(users)
data(words)
data_list <- list(train = train[1:1000, ], users = users, words = words)
caret_train(data_list)
}

