% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{cast_covariates}
\alias{cast_covariates}
\title{Create an array representing covariate information}
\usage{
cast_covariates(mZ, x_names = NULL)
}
\arguments{
\item{mZ}{A "long" matrix that can be automatically cast by data.table.
So, the first two columns should be row, column, and slice ids, the fourth
should be the covariate value to fill in.\}

\item{x_names}{An optional list of row names [character vector] to match
against. This can be used to ensure consistency in the rows across tables.
If a user appears in x-names but not in mZ, a new row of all NAs included.
By default, the cast version of mZ is returned without any reordering of
users.}

\item{mZ}{}
}
\value{
An array Z with the structure described above.
}
\description{
All covariate information is provided in "long" form. That is,
different user-artist-question pairs are rows of the "words" matrix. Instead,
we need an array whose elements z_ijk give the response that user i gave
about artist j for question k.
}

