% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval.R
\name{contiguous_folds}
\alias{contiguous_folds}
\title{Create folds of contiguous indices}
\usage{
contiguous_folds(n, K)
}
\arguments{
\item{n}{How many indices have to be split into folds?}

\item{K}{How many folds should be created?}
}
\description{
Create folds of contiguous indices
}

