% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval.R
\name{evaluate}
\alias{evaluate}
\title{Generic Evaluation Function}
\usage{
evaluate(data_list, train_fun, train_opts, pred_fun, K = 5, n_rep = 1)
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
  train: A data.frame giving the artist, track, user, rating, and time info. \cr
  words: A matrix giving word indicators for artist-user pairs. \cr
  users: A matrix givin gsurvey results for each user. \cr}

\item{train_fun}{A function taking arguments data_list and train_opts, and
returning a trained model that can be used for prediction.}

\item{train_opts}{A list giving options for how to train the model in
train_fun.}

\item{pred_fun}{A function that given the output of train_fun and a new test
version of data_list returns predictions for each row of the test data.}

\item{K}{The number of cv-folds in cross-validation.}

\item{n_rep}{The number of cv replicates to perform.}
}
\value{
A list with the following elements, \cr
   mean_rmse: The mean of the RMSEs calculated over reps and folds.
   err: A K x n_rep matrix, whose kj^th element gives the error when holding
   out the k^th fold in the j^th rep.
}
\description{
Generic Evaluation Function
}

