% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{expand_by_track}
\alias{expand_by_track}
\title{Copy covariates associated with artists to match tracks}
\usage{
expand_by_track(Z, track_names, artist_track_map)
}
\arguments{
\item{Z}{A user x artist x question covariates matrix.}

\item{track_names}{The columns of the ratings matrix on which we are doing
matrix completion. After applying this function, the column names of Z will
be this vector.}

\item{artist_track_map}{A two column data.table mapping artists (first col)
to associated tracks (second col).}
}
\value{
Z The same matrix Z as before, but with responses for individual
artists repeated to correspond with the track level labels.
}
\description{
Unfortunately, the covariates in the EMI data are only
associated with individual artists, not tracks, which is the level at which
we are doing matrix completion. So, we need to expand the covariates matrix Z
from the artist level to the track level, using the artist_track_map.
}

