% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_data_type}
\alias{get_data_type}
\title{Get matrix type}
\usage{
get_data_type(X)
}
\arguments{
\item{X}{The matrix to check.}
}
\value{
Either "users" or "words" depending on the matrix type.
}
\description{
Check whether the input matrix is the users or words matrix
}

