% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_desc}
\alias{get_desc}
\title{Get Descriptions}
\usage{
get_desc(X)
}
\value{
Y A data.table with the descriptions of each data type.
}
\description{
This will return columns that are not the numeric question
variables and not the keys. E.g., this returns gender, age, working, etc.
}

