% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval.R
\name{get_folds}
\alias{get_folds}
\title{Get CV Folds for evaluation}
\usage{
get_folds(n = 188690, K = 10, n_rep = 1, contiguous = FALSE)
}
\arguments{
\item{n}{How many indices have to be split into folds?}

\item{K}{How many folds should be created?}

\item{n_rep}{How many independent sets of cross-validation fold indices
should be returned?}

\item{contiguous}{Should the CV folds be of contiguous groups of indices?}
}
\description{
Wrapper for caret::createFolds with some slight generalizations.
}

