% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_keys}
\alias{get_keys}
\title{Get the keys for the matrix}
\usage{
get_keys(X)
}
\arguments{
\item{X}{The matrix to get keys for.}
}
\value{
Either the RESPID or RESPID + artists IDs for the users or words
matrices.
}
\description{
Get the keys associated with the users and words matrices.
}

