% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_numeric}
\alias{get_numeric}
\title{Get Quantitative Features}
\usage{
get_numeric(X)
}
\arguments{
\item{X}{The matrix to get numeric columns for.}
}
\value{
The matrix of quantitative features.
}
\description{
Get Quantitative Features
}

