% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{impute_median}
\alias{impute_median}
\title{Impute NAs in numeric columns of a data.frame with medians}
\usage{
impute_median(X)
}
\description{
Impute NAs in numeric columns of a data.frame with medians
}

