% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_gd.R
\name{lf_gd}
\alias{lf_gd}
\title{Regular gradient descent for latent factor model with covariates}
\usage{
lf_gd(X, k_factors = 3, lambdas = rep(1, 2), n_iter = 5, gamma = 0.1)
}
\arguments{
\item{X}{The ratings matrix. Unobserved entries must be marked NA. Users
must be along rows, and tracks must be along columns.}

\item{k_factors}{The number of latent factors for the problem.}

\item{lambdas}{The regularization parameters for P, Q, and beta, respectively.}

\item{n_iter}{The number of gradient descent iterations to run.}

\item{gamma}{The step-size in the gradient descent.}
}
\value{
A list with the following elements, \cr
  $P The learned user latent factors. \cr
  $Q The learned track latent factors.
  $beta The learned regression coefficients.
}
\description{
Regular gradient descent for latent factor model with covariates
}

