% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_gd_cov.R
\name{lf_gd_cov}
\alias{lf_gd_cov}
\title{Regular gradient descent for latent factor model with covariates}
\usage{
lf_gd_cov(X, Z, k_factors = 3, lambas = rep(1, 3), n_iter = 5,
  gamma = 0.1)
}
\arguments{
\item{X}{The ratings matrix. Unobserved entries must be marked NA. Users
must be along rows, and tracks must be along columns.}

\item{Z}{The covariates associated with each pair. This must be shaped as an
array with users along rows, tracks along columns, and features along
slices.}

\item{k_factors}{The number of latent factors for the problem.}

\item{n_iter}{The number of gradient descent iterations to run.}

\item{gamma}{The step-size in the gradient descent.}

\item{lambdas}{The regularization parameters for P, Q, and beta, respectively.}
}
\value{
A list with the following elements, \cr
  $P The learned user latent factors. \cr
  $Q The learned track latent factors.
  $beta The learned regression coefficients.
}
\description{
Regular gradient descent for latent factor model with covariates
}

