% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lf_sgd}
\alias{lf_sgd}
\title{Latent Factor modeling using Stochastic Gradient Descent}
\usage{
lf_sgd(R, obs_ix, lambdas, eta = 0.05, iter_max = 1000L, d = 5L)
}
\arguments{
\item{R}{A user x item matrix of ratings.}

\item{obs_ix}{A n_obs x 2 matrix whose rows give indices at which ratings
were observed.}

\item{lambdas}{A vector of regularization parameters, as in the objective
function given above.}

\item{eta}{The learning rate for the stochastic gradient descent.}

\item{iter_max}{The number of iterations to run the SGD.}

\item{d}{The dimension of the latent factors.}
}
\value{
The fitted parameters from the objective function below, \cr
   bu: Bias terms for each user. \cr
   bi: Bias terms for each item. \cr
   P: A user x d latent factor matrix for users \cr.
   Q: A item x d latent factor matrix for items.
}
\description{
Decompose a matrix R to minimize
    ||R - b_u * 1 + 1 * b_i' - PQ||_{F}^{2} + lambda_1 * ||b_{u}||_{2}^{2} +
       lambda_2 * ||b_i||_{2}^{2} + lambda_3 * ||P||_{F}^{2} +
       lambda_4 * ||Q||_{F}^{2}
}

