% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_sgd.R
\name{lf_sgd_predict}
\alias{lf_sgd_predict}
\title{Make predictions from the lf-sgd model}
\usage{
lf_sgd_predict(model, data_list)
}
\arguments{
\item{model}{The output of a call to lf_sgd_train().}

\item{data_list}{The new test samples to make predictions over.}
}
\value{
preds A vector of predictions on the new samples.
}
\description{
Make predictions from the lf-sgd model
}

