% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_sgd.R
\name{lf_sgd_train}
\alias{lf_sgd_train}
\title{Train a latent factor SGD model}
\usage{
lf_sgd_train(data_list, train_opts = list())
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
  train: A data.table giving the artist, track, user, rating, and time info. \cr
  words: A data.table giving word indicators for artist-user pairs. \cr
  users: A data.table giving survey results for each user. \cr
We don't ever actually use words or users in this model. This convention is
enforced for simplicity only.}

\item{train_opts}{A list with possible training options as specified in
merge_lf_sgd_opts().}
}
\value{
A latent factor model, with elements
   mu: The global mean of R. \cr
   bu: Bias terms for each user. \cr
   bi: Bias terms for each item. \cr
   P: A user x d latent factor matrix for users \cr.
   Q: A item x d latent factor matrix for items.
}
\description{
Train a latent factor SGD model
}

