% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_predictor.R
\name{mean_model}
\alias{mean_model}
\title{"Train" a model that predicts the means within groups.}
\usage{
mean_model(data_list, train_opts = list(type = "User"))
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
  train: A data.table giving the artist, track, user, rating, and time info. \cr
  words: A data.table giving word indicators for artist-user pairs. \cr
  users: A data.table giving survey results for each user. \cr
We don't ever actually use words or users in this model. This convention is
enforced for simplicity only.}

\item{train_opts}{A list with an element called type, specifying the group to
average over.}
}
\value{
A list with the following elements \cr
   cur_means: The means within groups for this model.
   type: The type of group over which we calculated means.
   global_mean: The mean over all ratings.
}
\description{
"Train" a model that predicts the means within groups.
}

