% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_predictor.R
\name{mean_pred}
\alias{mean_pred}
\title{"Predict" using means within groups}
\usage{
mean_pred(model, data_list)
}
\arguments{
\item{data_list}{A list in the same form as in the input to mean_model, but
with only the test cases.}

\item{mean_model}{The output of a call to mean_model.}
}
\value{
Predictions on the test cases.
}
\description{
"Predict" using means within groups
}

