% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_predictor.R
\name{mean_predictor}
\alias{mean_predictor}
\title{Predict using the mean of a group}
\usage{
mean_predictor(train, test, type = "User")
}
\arguments{
\item{train}{The training data, with a column for "type" and a column for
true ratings.}

\item{test}{The test data, with a column for "type", whose ratings we want to
predict.}

\item{type}{The column to take averages over. Defaults to user.}
}
\value{
pred_means The predicted ratings for the test data.
}
\description{
Predict using the mean of a group
}

