% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{merge_caret_opts}
\alias{merge_caret_opts}
\title{Default options for using generic predicton algorithms in caret}
\usage{
merge_caret_opts(opts = list())
}
\arguments{
\item{opts}{A partially specified list of options to use when caret models.
The currently supported options are
  $process_opts: A list of processing options to pass the feature matrix
   through, via the preprocess_data() function.
  $method: The method to use from caret. See https://topepo.github.io/caret/modelList.html
  $postprocess_fun: A function used to postprocess the predictions at test
   time.
  $train_control: The trControl argument to caret::train().}
}
\value{
The original opts list with defaults filled in.
}
\description{
Default options for using generic predicton algorithms in caret
}

