% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{merge_intersect_X}
\alias{merge_intersect_X}
\title{Wrapper for merging by all shared columns, but only keeping X}
\usage{
merge_intersect_X(X, Y)
}
\description{
Wrapper for merging by all shared columns, but only keeping X
}

