% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_sgd.R
\name{merge_lf_sgd_opts}
\alias{merge_lf_sgd_opts}
\title{Merge default options for the latent factor model trained with SGD}
\usage{
merge_lf_sgd_opts(opts = list())
}
\arguments{
\item{opts}{A partially specified list of options, with the following
possible elements, \cr
  lambdas: A vector of regularization parameters. \cr
  eta: The learning rate for the stochastic gradient descent. \cr
  iter__max: The number of iterations to run the SGD. \cr
  d: The dimension of the latent factors. \cr}
}
\value{
The opts list with defaults filled in.
}
\description{
Merge default options for the latent factor model trained with SGD
}

