% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{merge_preprocess_opts}
\alias{merge_preprocess_opts}
\title{Merge default prprocessing options}
\usage{
merge_preprocess_opts(opts = list())
}
\description{
Merge default prprocessing options
}

