% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_predict.R
\name{merge_svd_opts}
\alias{merge_svd_opts}
\title{Default options for SVD predictions}
\usage{
merge_svd_opts(opts = list())
}
\arguments{
\item{opts}{A partially specified list of options to use when performing the
SVD imputation. The currently supported options are
  $alpha The rescaling parameter to apply after taking the SVD [to account
   for the many 0's in the matrix on which we take the SVD]
  $k The rank of the imputation. Defaults to 10.
  $keep_cols The columns of the "tall" training data that will be "cast"
  $max_val The top value to truncate predictions at.
  $min_val THe bottom value to truncate predictions at.}
}
\description{
Default options for SVD predictions
}

