% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{merge_users_covariates}
\alias{merge_users_covariates}
\title{Merge in user specific covariates}
\usage{
merge_users_covariates(Z, users)
}
\arguments{
\item{Z}{An array with dimensions user x artist x question response.}

\item{users}{A matrix whose first column is a user id, and whose remaining
columns give descriptions of musical preference.}

\item{We}{would like to include users' descriptions of their personal tastes
in this regression problem. However, our regression array has the form
user x artist x question, so to merge it in, we need to copy the same
responses for every artist.}
}
\value{
Z_users The matrix Z with the questions dimension increased to
account for the additional covariates from the users matrix. If a user in Z
did not appear in the users matrix, the new question is filled in with all
NAs.
}
\description{
Merge in user specific covariates
}

