% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{plot_lasso_coef}
\alias{plot_lasso_coef}
\title{Alternative visualization of glmnet coefficients}
\usage{
plot_lasso_coef(beta)
}
\arguments{
\item{beta}{The coefficient matrix of a glmnet object.}
}
\value{
res_plot A heatmap of the lasso paths.
}
\description{
Visualize significant coefficients using a heatmap,
instead of a trace of the lasso paths.
}

