% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{postprocess_preds}
\alias{postprocess_preds}
\title{Extract predictions from X_hat matrix}
\usage{
postprocess_preds(X_hat, newdata)
}
\arguments{
\item{X_hat}{A user x track matrix with predictions.}

\item{newdata}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
 $train: A data.frame giving the artist, track, user, and time info. \cr
 $words: A matrix giving word indicators for artist-user pairs. \cr
 $users: A matrix givin gsurvey results for each user. \cr}
}
\value{
A vector of ratings corresponding to the user x track pairs in $train
from the newdata.
}
\description{
Extract predictions from X_hat matrix
}

