% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{prepare_covariates}
\alias{prepare_covariates}
\title{Wrapper to do all the covariates preparation.}
\usage{
prepare_covariates(users, words, x_names, track_names, artist_track_map)
}
\arguments{
\item{users}{A matrix whose first column is a user id, and whose remaining
columns give descriptions of musical preference.}

\item{words}{The words data.table in directly from the EMI challenge.}

\item{x_names}{An optional list of row names [character vector] to match
against. This can be used to ensure consistency in the rows across tables.
If a user appears in x-names but not in mZ, a new row of all NAs included.
By default, the cast version of mZ is returned without any reordering of
users.}

\item{track_names}{The columns of the ratings matrix on which we are doing
matrix completion. After applying this function, the column names of Z will
be this vector.}

\item{artist_track_map}{A two column data.table mapping artists (first col)
to associated tracks (second col).}
}
\value{
A covariates matrix with dimensions user x track x question. If no
response was recorded, the array element is filled with an NA.
}
\description{
Wrapper to do all the covariates preparation.
}

