% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{prepare_features}
\alias{prepare_features}
\title{Prepare features for generic caret training}
\usage{
prepare_features(data_list, process_opts)
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
  train: A data.frame giving the artist, track, user, rating, and time info. \cr
  words: A matrix giving word indicators for artist-user pairs. \cr
  users: A matrix givin gsurvey results for each user. \cr}

\item{process_opts}{A list of processing options to apply to the feature
matrix, see merge_process_opts().}
}
\value{
A lsit with the following elements, \cr
  $X: A feature matrix whose rows correspond to the user - track pairs
  $y: A response vector fo ratings associated to the rows of X
in the train data set.
}
\description{
Prepare features for generic caret training
}

