% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{prepare_pred_Z}
\alias{prepare_pred_Z}
\title{Merge train and test data, for imputation in Z}
\usage{
prepare_pred_Z(train_list, newdata, x_dimnames)
}
\arguments{
\item{train_list}{A data_list of the form
$train: A data.frame giving the artist, track, user, rating, and time info. \cr
$words: A matrix giving word indicators for artist-user pairs. \cr
$users: A matrix givin gsurvey results for each user. \cr}

\item{newdata}{A test data_list, of the same form as train_list.}

\item{x_dimnames}{A list of dimnames of the ratings matrix.}
}
\value{
Z A user x track x question covariates matrix.
}
\description{
Merge train and test data, for imputation in Z
}

