% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocessing}
\usage{
preprocess_data(X, opts = list())
}
\arguments{
\item{X}{A data.frame or data.table to preprocess.}

\item{opts}{A (potentially partially specified) list of preprocessing
options. See merge_preprocess_opts for possible options, and their defaults.}
}
\value{
X The processed version of X. Should have same dimensions as input.
}
\description{
Preprocessing
}

