% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svd_cov}
\alias{svd_cov}
\title{Regular gradient descent for latent factor model with covariates}
\usage{
svd_cov(X, Z_vec, k_factors, lambdas, n_iter, batch_samples, batch_factors,
  gamma_pq, gamma_beta, verbose)
}
\arguments{
\item{X}{The ratings matrix. Unobserved entries must be marked NA. Users
must be along rows, and tracks must be along columns.}

\item{k_factors}{The number of latent factors for the problem.}

\item{lambdas}{The regularization parameters for P, Q, and beta, respectively.}

\item{n_iter}{The number of gradient descent iterations to run.}

\item{batch_samples}{The proportion of the observed entries to sample in the
gradient for each factor in the gradient descent.}

\item{batch_factors}{The proportion of latent factors to update at each
iteration.}

\item{verbose}{Print the objective at each iteration of the descent?}

\item{Z}{The covariates associated with each pair. This must be shaped as an
array with users along rows, tracks along columns, and features along
slices.}

\item{gamma}{The step-size in the gradient descent.}
}
\value{
A list with the following elements, \cr
  $P The learned user latent factors. \cr
  $Q The learned track latent factors. \cr
  $beta The learned regression coefficients.
}
\description{
Regular gradient descent for latent factor model with covariates
}

