% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{svd_cov_impute}
\alias{svd_cov_impute}
\title{Perform SVD with Covariates imputation}
\usage{
svd_cov_impute(X, Z, opts)
}
\arguments{
\item{X}{A users x tracks matrix of ratings.}

\item{Z}{A users x tracks x question covariates matrix.}

\item{opts}{A list of options to use in prediction.}
}
\value{
A list with the following elements: \cr
 $X_hat A users x tracks matrix of predicted ratings. \cr
 $res The full output of the SVD imputation with covariates, svd_cov, in the
  src directory. /\cr
}
\description{
Perform SVD with Covariates imputation
}

