% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{svd_cov_predict}
\alias{svd_cov_predict}
\title{Matrix completion predictions using the SVD}
\usage{
svd_cov_predict(trained_model, newdata)
}
\arguments{
\item{trained_model}{The output of svd_train()}

\item{newdata}{A new data_list of the same form as data_list in the input to
svd_train()}
}
\value{
y_hat Predictions y_hat for every new user-track pair in the newdata.
}
\description{
Matrix completion predictions using the SVD
}

