% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{svd_cov_train}
\alias{svd_cov_train}
\title{Wrapper for saving information in SVD with covariates model}
\usage{
svd_cov_train(data_list, opts = list())
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
 $train: A data.frame giving the artist, track, user, rating, and time info. \cr
 $words: A matrix giving word indicators for artist-user pairs. \cr
 $users: A matrix givin gsurvey results for each user. \cr}

\item{opts}{A list of svd with covariates imputation options. See
merge_svd_cov_opts.}
}
\value{
A list that can be used for imputation with covariates. It contains
the following elements,
 $X A users x tracks matrix of ratings.
 $Z A users x tracks x question covariates matrix.
 $opts A list of options to use in prediction.
}
\description{
This is mainly useful so that we can use the cv evaluation
functions. It just saves the data and options, actual imputation is done
at prediction time.
}
\examples{
data(train)
data(users)
data(words)
data_list <- list(train = train[1:1000, ],
                  words = words,
                  users = users)
svd_cov_train(data_list)
}

