% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_predict.R
\name{svd_impute}
\alias{svd_impute}
\title{Perform core SVD imputation}
\usage{
svd_impute(R, k, alpha, min_val, max_val)
}
\arguments{
\item{R}{A users x ratings rating matrix, with NAs whenever a rating is
unknown.}

\item{k}{The rank of the SVD approximation to use.}

\item{alpha}{The rescaling factor after performing imputation, to account for
the many 0's in the originally SVD'd matrix.}

\item{min_val}{The minimum value to truncate predictions at.}

\item{max_val}{The maximum value to truncate predictions at.}
}
\value{
pR_hat The version of R with imputations filled in.
}
\description{
Perform core SVD imputation
}

