% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_predict.R
\name{svd_train}
\alias{svd_train}
\title{Wrapper saving information in SVD model}
\usage{
svd_train(data_list, opts = list())
}
\arguments{
\item{data_list}{A list containing the following elements, paralleling the
input provided by Kaggle,\cr
  train: A data.frame giving the artist, track, user, rating, and time info. \cr
  words: A matrix giving word indicators for artist-user pairs. \cr
  users: A matrix givin gsurvey results for each user. \cr}

\item{opts}{A list of svd imputation options. See merge_svd_opts.}
}
\description{
This is mainly useful so that we can use the cv evaluation
functions. It just saves the data and options, actual imputation is done
at prediction time.
}

