% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_gd_cov.R
\name{update_beta}
\alias{update_beta}
\title{Gradient step for regression coefficient}
\usage{
update_beta(X, Z, P, Q, beta, lambda, gamma)
}
\arguments{
\item{X}{The ratings matrix. Unobserved entries must be marked NA. Users
must be along rows, and tracks must be along columns.}

\item{P}{The learned user latent factors.}

\item{Q}{The learned track latent factors.}

\item{beta}{The learned regression coefficients.}

\item{lambda}{The regularization parameter for beta.}

\item{gamma}{The step-size in the gradient descent.}
}
\value{
The update regression coefficients.
}
\description{
Gradient step for regression coefficient
}

