% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_gd_cov.R
\name{update_factor}
\alias{update_factor}
\title{Gradient step for Latent Factors}
\usage{
update_factor(x, z, p_i, Q, beta, lambda, gamma)
}
\arguments{
\item{x}{Either a single row or single column of X, the ratings matrix.}

\item{z}{A single matrix from the covariates array, associated with the
either a single row or column of X.}

\item{p_i}{The latent factors associated with a single user or track.}

\item{Q}{The latent factors associated with either all tracks or all users,
depending on the type of latent factor in p_i.}

\item{beta}{The learned regression coefficients.}

\item{lambda}{The regularization parameter for the current latent factor.}

\item{gamma}{The step-size in the gradient descent.}
}
\value{
The update latent factor for the current user / track.
}
\description{
Gradient step for Latent Factors
}

