% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demographic_predictor.R
\name{user_info_covariates}
\alias{user_info_covariates}
\title{Build predictor matrix incorporating user information}
\usage{
user_info_covariates(train, users)
}
\description{
Build predictor matrix incorporating user information
}

