% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demographic_predictor.R
\name{user_info_predictor}
\alias{user_info_predictor}
\title{Predictor with user information}
\usage{
user_info_predictor(train, test, users, trControl = NULL)
}
\arguments{
\item{train}{}

\item{test}{}

\item{users}{}

\item{trControl}{}
}
\description{
Predictor with user information
}

