% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd_cov.R
\name{merge_svd_cov_opts}
\alias{merge_svd_cov_opts}
\title{Default options for SVD with covariates}
\usage{
merge_svd_cov_opts(opts = list())
}
\arguments{
\item{opts}{A partially specified list of options to use when performing the
SVD imputation. The currently supported options are}
}
\description{
Default options for SVD with covariates
}

